Revert "refactor: rename ingredientsDataAPI to stellarBurgersAPI"
This reverts commit 3016962432ca16d3fcff06fe928625173a8dbc3a.
