+++ src/services/ingredientsDataAPI.js
@@ -1,7 +1,7 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
 import baseUrl from '../utils/api/constants';
 
-export const stellarBurgersAPI = createApi({
+export const ingredientsDataAPI = createApi({
   baseQuery: fetchBaseQuery({ baseUrl }),
   endpoints: (build) => ({
     getIngredientsData: build.query({
@@ -20,4 +20,4 @@ export const stellarBurgersAPI = createApi({
   }),
 });
 
-export const { useGetIngredientsDataQuery, useGetOrderDataMutation } = stellarBurgersAPI;
+export const { useGetIngredientsDataQuery, useGetOrderDataMutation } = ingredientsDataAPI;
