+++ src/services/rootReducer.js
@@ -2,13 +2,13 @@ import { combineReducers } from 'redux';
 import activeIngredientDataReducer from './activeIngredientDataSlice';
 import orderDetailsDataReducer from './orderDetailsDataSlice';
 import burgerConstructorIngredientsDataReducer from './burgerConstructorIngredientsDataSlice';
-import { stellarBurgersAPI } from './stellarBurgersAPI';
+import { ingredientsDataAPI } from './ingredientsDataAPI';
 
 const rootReducer = combineReducers({
   activeIngredientDataReducer,
   orderDetailsDataReducer,
   burgerConstructorIngredientsDataReducer,
-  [stellarBurgersAPI.reducerPath]: stellarBurgersAPI.reducer,
+  [ingredientsDataAPI.reducerPath]: ingredientsDataAPI.reducer,
 });
 
 export default rootReducer;
