+++ src/services/store.js
@@ -1,10 +1,11 @@
 import { configureStore } from '@reduxjs/toolkit';
-import { stellarBurgersAPI } from './stellarBurgersAPI';
+import { ingredientsDataAPI } from './ingredientsDataAPI';
 import rootReducer from './rootReducer';
 
 const store = configureStore({
   reducer: rootReducer,
-  middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(stellarBurgersAPI.middleware),
+  middleware: (getDefaultMiddleware) =>
+    getDefaultMiddleware().concat(ingredientsDataAPI.middleware),
 });
 
 export default store;
